Add preliminary code for handling impressions
Impressions are first-class DB entities representing an impression one character formed
about another character based on a particular introspection action. Every impression has
a positive or negative numerical value associated with it, and the 'charge' from one
character toward another is now derived by summing the values of the individual impressions
that the first character has formed of the second. We store only a limited number of
impressions for each directed character->character edge at a time; right now, there's three
'slots' each for positive and negative impressions from one character to another.

Whenever we attempt to add a new impression from character A to character B, we first check
whether there's an open slot for a new impression of this valence between these characters.
If there is, we just unconditionally add the new impression. If there isn't, we check the
value of the new impression we're trying to add against the lowest-valued of the currently
stored same-valence impressions from A to B. If the new impression has a 'greater' value
(i.e. more positive for positive-valence impressions, more negative for negative-valence
impressions) than the 'weakest' of the currently stored same-valence impressions for this
directed character pair, we displace the 'weakest' existing impression (dropping it from the
database entirely) and add the new impression in its place. Then we recalculate the charge
from A to B based on the impressions that are now present in the database.

The net effect is that characters start out with very fuzzy or 'weak' impressions of one
another, then gradually converge on stronger and stronger impressions over time. Small
slights and one-off compliments are easily 'forgotten', while larger betrayals and
significant gestures of kindness tend to stick around as long-term defining aspects of a
relationship. In the future, I want to create actions that permit characters to communicate
impressions to one another, as well as to consolidate many similar 'weak' impressions into a
single 'stronger' impression.